rename: 컴포넌트 파일 폴더 추가
